Demonstration of how to grab huffman data from a jpg
This quick demonstration sequentially reads a jpg file while listening for the huffman data indicator 0xFFC4, then prints out all of the huffman data in hex. Other pieces of data in the jpg file could probably be displayed in the same way by listening for a different indicator. To try this demo, point the FileInputStream at any jpg file.